+++ ivy/functional/backends/jax/experimental/statistical.py
@@ -2,9 +2,11 @@ import jax.numpy as jnp
 from typing import Optional, Union, Tuple, Sequence
 
 from ivy.functional.backends.jax import JaxArray
+import jax.lax as jlax
 import ivy
 from ivy.func_wrapper import with_unsupported_dtypes
 from . import backend_version
+from ..statistical import _infer_dtype
 
 
 @with_unsupported_dtypes(
@@ -264,3 +266,124 @@ def cov(
         fweights=fweights,
         aweights=aweights,
     )
+
+
+def cummax(
+    x: JaxArray,
+    /,
+    *,
+    axis: int = 0,
+    exclusive: bool = False,
+    reverse: bool = False,
+    dtype: Optional[jnp.dtype] = None,
+    out: Optional[JaxArray] = None,
+) -> Tuple[JaxArray, JaxArray]:
+    if x.dtype in (jnp.bool_, jnp.float16):
+        x = x.astype(jnp.float64)
+    elif x.dtype in (jnp.int16, jnp.int8, jnp.uint8):
+        x = x.astype(jnp.int64)
+    elif x.dtype in (jnp.complex128, jnp.complex64):
+        x = jnp.real(x).astype(jnp.float64)
+
+    if exclusive or (reverse and exclusive):
+        if exclusive and reverse:
+            indices = __find_cummax_indices(jnp.flip(x, axis=axis), axis=axis)
+            x = jlax.cummax(jnp.flip(x, axis=axis), axis=axis)
+            x, indices = jnp.swapaxes(x, axis, -1), jnp.swapaxes(indices, axis, -1)
+            x, indices = jnp.concatenate(
+                (jnp.zeros_like(x[..., -1:]), x[..., :-1]), -1
+            ), jnp.concatenate(
+                (jnp.zeros_like(indices[..., -1:]), indices[..., :-1]), -1
+            )
+            x, indices = jnp.swapaxes(x, axis, -1), jnp.swapaxes(indices, axis, -1)
+            res, indices = jnp.flip(x, axis=axis), jnp.flip(indices, axis=axis)
+        elif exclusive:
+            x = jnp.swapaxes(x, axis, -1)
+            x = jnp.concatenate((jnp.zeros_like(x[..., -1:]), x[..., :-1]), -1)
+            x = jnp.swapaxes(x, axis, -1)
+            indices = __find_cummax_indices(x, axis=axis)
+            res = jlax.cummax(x, axis=axis)
+        return res, indices
+
+    if reverse:
+        y = jnp.flip(x, axis=axis)
+        indices = __find_cummax_indices(y, axis=axis)
+        indices = jnp.flip(indices, axis=axis)
+    else:
+        indices = __find_cummax_indices(x, axis=axis)
+    return jlax.cummax(x, axis, reverse=reverse), indices
+
+
+def __find_cummax_indices(
+    x: JaxArray,
+    axis: int = 0,
+) -> JaxArray:
+    n, indice, indices = 0, [], []
+
+    if isinstance(x[0], JaxArray) and len(x[0].shape) >= 1:
+        if axis >= 1:
+            for ret1 in x:
+                indice = __find_cummax_indices(ret1, axis=axis - 1)
+                indices.append(indice)
+        else:
+            z_list = __get_index(x.tolist())
+            indices, n1 = x.copy(), {}
+            indices = jnp.zeros(jnp.asarray(indices.shape), dtype=x.dtype)
+            z_list = sorted(z_list, key=lambda i: i[1])
+            for y, y_index in z_list:
+                multi_index = y_index
+                if tuple(multi_index[1:]) not in n1:
+                    n1[tuple(multi_index[1:])] = multi_index[0]
+                    indices = indices.at[y_index].set(multi_index[0])
+                elif (
+                    y >= x[tuple([n1[tuple(multi_index[1:])]] + list(multi_index[1:]))]
+                ):
+                    n1[tuple(multi_index[1:])] = multi_index[0]
+                    indices = indices.at[y_index].set(multi_index[0])
+                else:
+                    indices = indices.at[y_index].set(n1[tuple(multi_index[1:])])
+    else:
+        n, indices = 0, []
+        for idx, y in enumerate(x):
+            if idx == 0 or x[n] <= y:
+                n = idx
+            indices.append(n)
+
+    return jnp.asarray(indices, dtype="int64")
+
+
+def __get_index(lst, indices=None, prefix=None):
+    if indices is None:
+        indices = []
+    if prefix is None:
+        prefix = []
+
+    if isinstance(lst, list):
+        for i, sub_lst in enumerate(lst):
+            sub_indices = prefix + [i]
+            __get_index(sub_lst, indices, sub_indices)
+    else:
+        indices.append((lst, tuple(prefix)))
+    return indices
+
+
+@with_unsupported_dtypes({"0.4.13 and below": "bfloat16"}, backend_version)
+def cummin(
+    x: JaxArray,
+    /,
+    *,
+    axis: int = 0,
+    exclusive: bool = False,
+    reverse: bool = False,
+    dtype: Optional[jnp.dtype] = None,
+    out: Optional[JaxArray] = None,
+) -> JaxArray:
+    if axis < 0:
+        axis = axis + len(x.shape)
+    dtype = ivy.as_native_dtype(dtype)
+    if dtype is None:
+        if dtype is jnp.bool_:
+            dtype = ivy.default_int_dtype(as_native=True)
+        else:
+            dtype = _infer_dtype(x.dtype)
+    return jlax.cummin(x, axis, reverse=reverse).astype(dtype)
