+++ ivy/functional/backends/paddle/experimental/statistical.py
@@ -2,6 +2,7 @@
 from typing import Optional, Union, Tuple, Sequence
 import paddle
 import ivy.functional.backends.paddle as paddle_backend
+import ivy
 
 # local
 from ivy.func_wrapper import with_unsupported_device_and_dtypes
@@ -451,3 +452,163 @@ def cov(
     return paddle.linalg.cov(
         X, rowvar=rowVar, ddof=ddof, fweights=fweights, aweights=aweights
     )
+
+
+@with_unsupported_device_and_dtypes(
+    {"2.5.0 and below": {"cpu": ("uint16", "bfloat16")}}, backend_version
+)
+def cummax(
+    x: paddle.Tensor,
+    /,
+    *,
+    axis: int = 0,
+    exclusive: bool = False,
+    reverse: bool = False,
+    dtype: Optional[paddle.dtype] = None,
+    out: Optional[paddle.Tensor] = None,
+) -> Tuple[paddle.Tensor, paddle.Tensor]:
+    if x.dtype in (paddle.bool, paddle.float16):
+        x = paddle.cast(x, "float64")
+    elif x.dtype in (paddle.int16, paddle.int8, paddle.uint8):
+        x = paddle.cast(x, "int64")
+    elif x.dtype in (paddle.complex128, paddle.complex64):
+        x = paddle.cast(paddle.real(x), "float64")
+
+    if not (exclusive or reverse):
+        return __find_cummax(x, axis=axis)
+
+    elif exclusive and reverse:
+        x, indices = __find_cummax(ivy.flip(x, axis=(axis,)), axis=axis)
+        x, indices = ivy.swapaxes(x, axis, -1), ivy.swapaxes(indices, axis, -1)
+        x = ivy.concat((ivy.zeros_like(x[..., -1:]), x[..., :-1]), axis=-1)
+        indices = ivy.concat(
+            (ivy.zeros_like(indices[..., -1:]), indices[..., :-1]), axis=-1
+        )
+        x, indices = ivy.swapaxes(x, axis, -1), ivy.swapaxes(indices, axis, -1)
+        return ivy.flip(x, axis=(axis,)), ivy.flip(indices, axis=(axis,))
+
+    elif exclusive:
+        x = ivy.swapaxes(x, axis, -1)
+        x = ivy.concat((ivy.zeros_like(x[..., -1:]), x[..., :-1]), axis=-1)
+        x = ivy.swapaxes(x, axis, -1)
+        x, indices = __find_cummax(x, axis=axis)
+
+        return x, indices
+
+    else:
+        x, indices = __find_cummax(ivy.flip(x, axis=(axis,)), axis=axis)
+        return ivy.flip(x, axis=axis), ivy.flip(indices, axis=axis)
+
+
+def __find_cummax(
+    x: paddle.Tensor, axis: int = 0, dtype: Optional[paddle.dtype] = None
+) -> Tuple[paddle.Tensor, paddle.Tensor]:
+    indices = []
+    values = []
+    x_dtype = x.dtype if dtype is None else dtype
+    if (
+        isinstance(x.tolist()[0], list)
+        and len(x[0].shape) >= 1
+        and (
+            (type(x[0]) == paddle.Tensor)
+            or (type(x[0]) == ivy.data_classes.array.array.Array)
+        )
+    ):
+        if axis >= 1:
+            if not isinstance(x, list):
+                x = x.tolist()
+            for ret1 in x:
+                value, indice = __find_cummax(
+                    paddle.to_tensor(ret1, dtype=x_dtype), axis=axis - 1, dtype=x_dtype
+                )
+                indices.append(indice)
+                values.append(value)
+        else:
+            x_list = x.numpy()
+            z_list = __get_index(x_list.tolist())
+            indices, values, n1 = x_list.copy(), x_list.copy(), {}
+            indices.fill(0)
+            values.fill(0)
+            z_list = sorted(z_list, key=lambda i: i[1])
+            for y, y_index in z_list:
+                multi_index = y_index
+                if tuple(multi_index[1:]) not in n1:
+                    n1[tuple(multi_index[1:])] = multi_index[0]
+                    indices[y_index] = multi_index[0]
+                    values[y_index] = y
+                elif (
+                    y
+                    >= x_list[
+                        tuple([n1[tuple(multi_index[1:])]] + list(multi_index[1:]))
+                    ]
+                ):
+                    n1[tuple(multi_index[1:])] = multi_index[0]
+                    indices[y_index] = multi_index[0]
+                    values[y_index] = y
+                else:
+                    indices[y_index] = n1[tuple(multi_index[1:])]
+                    values[y_index] = x_list[
+                        tuple([n1[tuple(multi_index[1:])]] + list(multi_index[1:]))
+                    ]
+    else:
+        if not isinstance(x, list):
+            x = x.tolist()
+        n = 0
+        for idx, y in enumerate(x):
+            if x[n] > y:
+                values.append(x[n])
+            elif x[n] <= y or idx == 0:
+                n = idx
+                values.append(y)
+            indices.append(n)
+
+    if type(x) == paddle.Tensor:
+        return paddle.to_tensor(values, dtype=x.dtype), paddle.to_tensor(
+            indices, dtype="int64"
+        )
+    else:
+        return ivy.array(values, dtype=x_dtype), ivy.array(indices, dtype="int64")
+
+
+def __get_index(lst, indices=None, prefix=None):
+    if indices is None:
+        indices = []
+    if prefix is None:
+        prefix = []
+
+    if isinstance(lst, list):
+        for i, sub_lst in enumerate(lst):
+            sub_indices = prefix + [i]
+            __get_index(sub_lst, indices, sub_indices)
+    else:
+        indices.append((lst, tuple(prefix)))
+    return indices
+
+
+@with_unsupported_device_and_dtypes(
+    {"2.5.0 and below": {"cpu": ("uint8", "int8", "int16")}},
+    backend_version,
+)
+def cummin(
+    x: paddle.Tensor,
+    /,
+    *,
+    axis: int = 0,
+    exclusive: bool = False,
+    reverse: bool = False,
+    dtype: Optional[paddle.dtype] = None,
+    out: Optional[paddle.Tensor] = None,
+) -> paddle.Tensor:
+    dtype = dtype if dtype is not None else x.dtype
+    if reverse:
+        x = paddle.flip(x, axis=[axis])
+    x_unstacked = paddle.unbind(x, axis=axis)
+    cummin_x_unstacked = []
+    cummin_x_unstacked.append(x_unstacked[0])
+    for i, x_sub in enumerate(x_unstacked[1:]):
+        cummin_x_sub = paddle.minimum(cummin_x_unstacked[i], x_sub)
+        cummin_x_unstacked.append(cummin_x_sub)
+    cummin_x = paddle.stack(cummin_x_unstacked, axis=axis)
+    if reverse:
+        cummin_x = paddle.flip(cummin_x, axis=[axis])
+    return cummin_x.cast(dtype)
