+++ ivy/functional/ivy/statistical.py
@@ -1146,240 +1146,6 @@ def cumprod(
     )
 
 
-@handle_array_function
-@to_native_arrays_and_back
-@handle_out_argument
-@handle_array_like_without_promotion
-@handle_nestable
-@handle_exceptions
-def cummax(
-    x: Union[ivy.Array, ivy.NativeArray],
-    /,
-    *,
-    axis: int = 0,
-    exclusive: bool = False,
-    reverse: bool = False,
-    out: Optional[ivy.Array] = None,
-) -> ivy.Array:
-    """
-    Return a tuple containing the cumulative maximum of elements of input along the
-    given axis and index location of each maximum value found along the given axis.
-
-    Parameters
-    ----------
-    x
-        Input array.
-    axis
-        Axis along which the cumulative maximum is computed. Default is ``0``.
-    exclusive
-        Whether to perform cummax exclusively. Default is ``False``.
-    reverse
-        Whether to perform the cummax from last to first element in the selected
-        axis. Default is ``False`` (from first to last element)
-    out
-        Optional output array, for writing the result to. It must have a shape that the
-        inputs broadcast to.
-
-    Returns
-    -------
-    ret
-        Array which holds the result of applying cummax at each
-        original array elements along the specified axis.
-
-    Examples
-    --------
-    With :class:`ivy.Array` input:
-
-    >>> x = ivy.array([-86, -19, 41, 88, -5, 80, 32, 87, -90, -12])
-    >>> y = ivy.cummax(x, exclusive=False, reverse=False)
-    >>> print(y)
-    [ivy.array([-86, -19, 41, 88, 88, 88, 88, 88, 88, 88]),
-      ivy.array([0, 1, 2, 3, 3, 3, 3, 3, 3, 3])]
-
-    >>> x = ivy.array([ 14,  15,  49, -24, -39])
-    >>> y = ivy.cummax(x, axis=0, exclusive=False, reverse=False)
-    >>> print(y)
-    [ivy.array([14, 15, 49, 49, 49]), ivy.array([0, 1, 2, 2, 2])]
-
-    >>> x = ivy.array([[ 63,  43, -16,  -4],[ 21,  82,  59,  33]])
-    >>> ivy.cummax(x, axis=0, reverse=False, dtype='int64', out=x)
-    >>> print(x)
-    [ivy.array([[ 63,  43, -16,  -4], [ 63,  82,  59,  33]]),
-    ivy.array([[0, 0, 0, 0],[0, 1, 1, 1]])]
-
-    >>> x = ivy.array([[-36,  83, -81],
-    ...                [ 23,  29,  63],
-    ...                [-83,  85,   2],
-    ...                [ 31,  25, -86],
-    ...                [-10, -52,   0],
-    ...                [ 22,  38,  55],
-    ...                [ 33,  54, -16]])
-    >>> y = ivy.cummax(x, axis=1, exclusive=True, reverse=False)
-    >>> print(y)
-    [ivy.array([[ 0,  0, 83],
-                [ 0, 23, 29],
-                [ 0,  0, 85],
-                [ 0, 31, 31],
-                [ 0,  0,  0],
-                [ 0, 22, 38],
-                [ 0, 33, 54]]), ivy.array([[0, 0, 2],
-                [0, 1, 2],
-                [0, 0, 2],
-                [0, 1, 1],
-                [0, 0, 0],
-                [0, 1, 2],
-                [0, 1, 2]])]
-
-    >>> x = ivy.array([73, 15, 47])
-    >>> ivy.cummax(x, axis=0, reverse=True, exclusive=True)
-    >>> print(y)
-    [ivy.array([47, 47,  0]), ivy.array([0, 0, 0])]
-
-    >>> x = ivy.array([-47, -14, -67, 15, -23, -45])
-    >>> ivy.cummax(x, axis=0, reverse=True, exclusive=False)
-    >>> print(y)
-    [ivy.array([ 15, 15, 15, 15, -23, -45]), ivy.array([2, 2, 2, 2, 1, 0])]
-    """
-    return current_backend(x).cummax(
-        x, reverse=reverse, axis=axis, exclusive=exclusive, out=out
-    )
-
-
-@handle_array_function
-@to_native_arrays_and_back
-@handle_out_argument
-@handle_array_like_without_promotion
-@handle_nestable
-@handle_exceptions
-def cummin(
-    x: Union[ivy.Array, ivy.NativeArray],
-    /,
-    *,
-    axis: int = 0,
-    reverse: bool = False,
-    dtype: Optional[Union[ivy.Dtype, ivy.NativeDtype]] = None,
-    out: Optional[ivy.Array] = None,
-) -> ivy.Array:
-    """
-    Return the cumulative minimum of the elements along a given axis.
-
-    Parameters
-    ----------
-    x
-        Input array.
-    axis
-        Axis along which the cumulative minimum is computed. Default is ``0``.
-    reverse
-        Whether to perform the cummin from last to first element in the selected
-        axis. Default is ``False`` (from first to last element)
-    dtype
-        Data type of the returned array. Default is ``None``.
-        If None, if the default data type corresponding to the data type “kind”
-        (integer or floating-point) of x has a smaller range of values than the
-        data type of x (e.g., x has data type int64 and the default data type
-        is int32, or x has data type uint64 and the default data type is int64),
-        the returned array must have the same data type as x.
-        If x has a floating-point data type, the returned array must have the
-        default floating-point data type.
-        If x has a signed integer data type (e.g., int16), the returned array
-        must have the default integer data type.
-        If x has an unsigned integer data type (e.g., uint16), the returned
-        array must have an unsigned integer data type having the same number of
-        bits as the default integer data type (e.g., if the default integer data
-        type is int32, the returned array must have a uint32 data type).
-        If the data type (either specified or resolved) differs from the data type
-        of x, the input array should be cast to the specified data type before
-        computing the product.
-    out
-        Optional output array, for writing the result to. It must have a shape that the
-        inputs broadcast to.
-
-    Returns
-    -------
-    ret
-        Array which holds the result of applying cummin at each
-        original array elements along the specified axis.
-
-    Examples
-    --------
-    With :class:`ivy.Array` input:
-
-    >>> x = ivy.array([1, 5, 2, 0])
-    >>> y = ivy.cummin(x)
-    >>> print(y)
-    ivy.array([1, 1, 1, 0])
-    >>> x = ivy.array([[6, 4, 2],
-    ...                [1, 3, 0]])
-    >>> y = ivy.zeros((2,3))
-    >>> ivy.cummin(x, axis=0, reverse=True, out=y)
-    >>> print(y)
-    ivy.array([[1., 3., 0.],
-        [1., 3., 0.]])
-
-    >>> x = ivy.array([[2, 4, 5],
-    ...                [3, 6, 5],
-    ...                [1, 3, 10]])
-    >>> ivy.cummin(x,axis=1,reverse=True, dtype='int64', out=x)
-    >>> print(x)
-    ivy.array([[ 2,  4,  5],
-        [ 3,  5,  5],
-        [ 1,  3, 10]])
-
-    With :class:`ivy.Container` input:
-
-    >>> x = ivy.Container(a=ivy.array([[1, 3, 5]]),
-    ...                   b=ivy.array([[3, 5, 7]]))
-    >>> y = ivy.cummin(x, axis= 0)
-    >>> print(y)
-    {
-        a: ivy.array([[1, 3, 5]]),
-        b: ivy.array([[3, 5, 7]])
-    }
-
-    >>> x = ivy.Container(a=ivy.array([[1, 3, 4]]),
-    ...                   b=ivy.array([[3, 5, 8],
-    ...                                [5, 6, 5]]),
-    ...                   c=ivy.array([[2, 4, 1],
-    ...                                [3, 6, 9],
-    ...                                [0, 2, 3]]))
-    >>> y = ivy.Container(a = ivy.zeros((1, 3)),
-    ...                   b = ivy.zeros((2, 3)),
-    ...                   c = ivy.zeros((3,3)))
-    >>> ivy.cummin(x,axis=1,reverse=True, out=y)
-    >>> print(y)
-    {
-    a: ivy.array([[1., 3., 4.]]),
-    b: ivy.array([[3., 5., 8.],
-                    [5., 5., 5.]]),
-    c: ivy.array([[1., 1., 1.],
-                    [3., 6., 9.],
-                    [0., 2., 3.]])
-    }
-
-    >>> x = ivy.Container(a=ivy.array([[0],[5]]),
-    ...                                [5]]),
-    ...                   b=ivy.array([[6, 8, 7],
-    ...                                [4, 2, 3]]),
-    ...                   c=ivy.array([[1, 2],
-    ...                                [3, 4],
-    ...                                [6, 4]]))
-    >>> ivy.cummin(x,axis=0,out=x)
-    >>> print(x)
-    {
-    a: ivy.array([[0],
-                    [0]]),
-    b: ivy.array([[6, 8, 7],
-                    [4, 2, 3]]),
-    c: ivy.array([[1, 2],
-                    [1, 2],
-                    [1, 2]])
-    }
-    """
-    return current_backend(x).cummin(
-        x, axis=axis, reverse=reverse, dtype=dtype, out=out
-    )
-
-
 @handle_exceptions
 @handle_nestable
 @handle_array_like_without_promotion
