+++ ivy_tests/test_ivy/test_functional/test_core/test_statistical.py
@@ -350,54 +350,6 @@ def test_cumsum(
     )
 
 
-@handle_test(
-    fn_tree="functional.ivy.cummax",
-    dtype_x_axis_castable=helpers.dtype_values_axis(
-        available_dtypes=helpers.get_dtypes("valid"),
-        min_num_dims=1,
-        max_num_dims=6,
-        min_value=-100,
-        max_value=100,
-        valid_axis=True,
-        allow_neg_axes=False,
-        max_axes_size=1,
-        force_int_axis=True,
-    ),
-    exclusive=st.booleans(),
-    reverse=st.booleans(),
-)
-def test_cummax(
-    *,
-    dtype_x_axis_castable,
-    exclusive,
-    reverse,
-    test_flags,
-    backend_fw,
-    fn_name,
-    on_device,
-    ground_truth_backend,
-):
-    input_dtype, x, axis = dtype_x_axis_castable
-    if "torch" in backend_fw.__name__:
-        assume(not test_flags.as_variable[0])
-        assume(not test_flags.test_gradients)
-
-    helpers.test_function(
-        ground_truth_backend=ground_truth_backend,
-        input_dtypes=input_dtype,
-        test_flags=test_flags,
-        fw=backend_fw,
-        fn_name=fn_name,
-        on_device=on_device,
-        x=x[0],
-        axis=axis,
-        exclusive=exclusive,
-        reverse=reverse,
-        rtol_=1e-1,
-        atol_=1e-1,
-    )
-
-
 # cumprod
 @handle_test(
     fn_tree="functional.ivy.cumprod",
@@ -443,44 +395,6 @@ def test_cumprod(
         dtype=castable_dtype,
         rtol_=1e-1,
         atol_=1e-1,
-    )
-
-
-# cummin
-@handle_test(
-    fn_tree="functional.ivy.cummin",
-    dtype_x_axis_castable=_get_castable_dtype(),
-    reverse=st.booleans(),
-)
-def test_cummin(
-    *,
-    dtype_x_axis_castable,
-    reverse,
-    test_flags,
-    backend_fw,
-    fn_name,
-    on_device,
-    ground_truth_backend,
-):
-    input_dtype, x, axis, castable_dtype = dtype_x_axis_castable
-    # ToDo: set as_variable_flags as the parameter generated by test_cummin once
-    # this issue is marked as completed https://github.com/pytorch/pytorch/issues/75733
-    if "torch" in backend_fw.__name__:
-        assume(not test_flags.as_variable[0])
-        assume(not test_flags.test_gradients)
-    helpers.test_function(
-        ground_truth_backend=ground_truth_backend,
-        input_dtypes=[input_dtype],
-        test_flags=test_flags,
-        fw=backend_fw,
-        fn_name=fn_name,
-        on_device=on_device,
-        x=x[0],
-        axis=axis,
-        reverse=reverse,
-        dtype=castable_dtype,
-        rtol_=1e-4,
-        atol_=1e-4,
     )
